files/requirements_fail.txt
# Libraries
aiohttp>=3.9.0
python-socketio==5.11.2

# TESTING
tenacity